Minor UI Changes

+++ app/src/main/java/za/co/woolworths/financial/services/android/models/dto/ColourSKUsPrices.kt
@@ -3,4 +3,4 @@ package za.co.woolworths.financial.services.android.models.dto
 /**
  * Created by Kunal Uttarwar on 15/02/24.
  */
-data class ColourSKUsPrices(val priceMin: Int, val priceMax: Int, val onSale: Boolean, val priceRange: Boolean)
+data class ColourSKUsPrices(val priceMin: Double, val priceMax: Double, val onSale: Boolean, val priceRange: Boolean)
